extracting out more accel from vm.rs
