+++ crates/runmat-vm/src/accel/fusion.rs
@@ -1,4 +1,113 @@
-//! Native acceleration execution and fusion runtime hooks.
-//!
-//! Port target:
-//! - fusion planning and execution helpers from `runmat-ignition/src/vm.rs`
+use crate::bytecode::Instr;
+use runmat_accelerate::InstrSpan;
+use runmat_builtins::Value;
+
+#[inline]
+pub fn value_kind(value: &Value) -> &'static str {
+    match value {
+        Value::Int(_) => "Int",
+        Value::Num(_) => "Num",
+        Value::Complex(_, _) => "Complex",
+        Value::Bool(_) => "Bool",
+        Value::LogicalArray(_) => "LogicalArray",
+        Value::String(_) => "String",
+        Value::StringArray(_) => "StringArray",
+        Value::CharArray(_) => "CharArray",
+        Value::Tensor(_) => "Tensor",
+        Value::ComplexTensor(_) => "ComplexTensor",
+        Value::Cell(_) => "Cell",
+        Value::Struct(_) => "Struct",
+        Value::GpuTensor(_) => "GpuTensor",
+        Value::Object(_) => "Object",
+        Value::HandleObject(_) => "HandleObject",
+        Value::Listener(_) => "Listener",
+        Value::FunctionHandle(_) => "FunctionHandle",
+        Value::Closure(_) => "Closure",
+        Value::ClassRef(_) => "ClassRef",
+        Value::MException(_) => "MException",
+        Value::OutputList(_) => "OutputList",
+    }
+}
+
+#[inline]
+pub fn summarize_value(i: usize, v: &Value) -> String {
+    match v {
+        Value::GpuTensor(h) => format!("in#{i}:GpuTensor shape={:?}", h.shape),
+        Value::Tensor(t) => format!("in#{i}:Tensor shape={:?}", t.shape),
+        Value::Num(n) => format!("in#{i}:Num({n:.6})"),
+        Value::Int(n) => format!("in#{i}:Int({})", n.to_i64()),
+        Value::Bool(b) => format!("in#{i}:Bool({})", if *b { 1 } else { 0 }),
+        Value::String(s) => format!("in#{i}:String({})", s),
+        _ => format!("in#{i}:{}", value_kind(v)),
+    }
+}
+
+pub fn fusion_span_live_result_count(instructions: &[Instr], span: &InstrSpan) -> Option<usize> {
+    if span.start > span.end || span.end >= instructions.len() {
+        return None;
+    }
+    let mut current_depth = 0usize;
+    for instr in &instructions[span.start..=span.end] {
+        let effect = instr.stack_effect()?;
+        if current_depth < effect.pops {
+            current_depth = effect.pops;
+        }
+        current_depth = current_depth - effect.pops + effect.pushes;
+    }
+    Some(current_depth)
+}
+
+pub fn fusion_span_has_vm_barrier(instructions: &[Instr], span: &InstrSpan) -> bool {
+    if span.start > span.end || span.end >= instructions.len() {
+        return true;
+    }
+    for instr in &instructions[span.start..=span.end] {
+        if matches!(
+            instr,
+            Instr::StoreIndex(_)
+                | Instr::StoreSlice(_, _, _, _)
+                | Instr::StoreSliceExpr { .. }
+                | Instr::StoreIndexCell(_)
+                | Instr::StoreMember(_)
+                | Instr::StoreMemberOrInit(_)
+                | Instr::StoreMemberDynamic
+                | Instr::StoreMemberDynamicOrInit
+        ) {
+            return true;
+        }
+    }
+    fusion_span_live_result_count(instructions, span) != Some(1)
+}
+
+pub struct StackSliceGuard<'a> {
+    stack: *mut Vec<Value>,
+    slice: Option<Vec<Value>>,
+    _marker: std::marker::PhantomData<&'a mut Vec<Value>>,
+}
+
+impl<'a> StackSliceGuard<'a> {
+    pub fn new(stack: &'a mut Vec<Value>, slice_start: usize) -> Self {
+        let slice = stack.split_off(slice_start);
+        Self {
+            stack,
+            slice: Some(slice),
+            _marker: std::marker::PhantomData,
+        }
+    }
+
+    pub fn slice(&self) -> &[Value] {
+        self.slice.as_ref().expect("stack slice missing").as_slice()
+    }
+
+    pub fn commit(mut self) {
+        self.slice = None;
+    }
+}
+
+impl Drop for StackSliceGuard<'_> {
+    fn drop(&mut self) {
+        if let Some(slice) = self.slice.take() {
+            unsafe { (&mut *self.stack).extend(slice) }
+        }
+    }
+}
